pyproject.toml
[project]
name = "mebot"
version = "0.1.0"
description = "Interactive CV chatbot for Ángel Nácar Jiménez using a multi-agent LLM pipeline"
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.12.0",
    "openai>=1.0.0",
    "jinja2>=3.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
    "jsonschema>=4.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["mebot*"]
